requirements.txt
pyserial==3.5
pandas==2.0.3
slack_sdk==3.21.3
tomli==2.0.1